requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-toastr==0.5.0
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
finvizfinance==0.14.6
yfinance==0.2.18
redis==4.6.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
matplotlib==3.7.2
schedule==1.2.0
scipy==1.11.1
cvxpy==1.3.2
scikit-learn==1.3.0
seaborn==0.12.2
aiohttp==3.8.5
redis==4.6.0
python-dotenv==1.0.0

# AI/ML Dependencies (for future implementation)
# transformers==4.30.0
# torch==2.0.1
# scikit-learn==1.3.0
# xgboost==1.7.6
# lightgbm==4.0.0
# pandas-ta==0.3.14b0
# yfinance==0.2.18
# praw==7.7.1
# tweepy==4.14.0
# newsapi-python==0.2.6